mocksvcs/mock_oidc/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic-settings>=2.7.0,<3.0.0
cryptography>=44.0.0,<45.0.0
PyJWT>=2.10.0,<3.0.0
python-multipart>=0.0.20,<1.0.0
